useGetSubmissionEvents
